Revert "Frontend (#148)"
This reverts commit ce7a61ce3318e73456a1cce526e2ec10f676a90b.
