--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "@inrupt/vocab-common-rdf": "^1.0.3"
-  }
-}
