+++ webapp/package.json
@@ -5,12 +5,9 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@inrupt/lit-generated-vocab-common": "^0.3.11",
-    "@inrupt/solid-client": "^1.21.0",
     "@inrupt/solid-ui-react": "^2.7.0",
-    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.2.5",
     "@mui/material": "^5.2.5",
-    "@mui/x-data-grid": "^5.7.0",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
@@ -25,7 +22,6 @@
     "react-hook-form": "^7.27.1",
     "react-router-dom": "^6.2.2",
     "typescript": "^4.5.4",
-    "uuid": "^8.3.2",
     "web-vitals": "^2.1.2"
   },
   "scripts": {
