+++ webapp/src/api/api.ts
@@ -1,4 +1,4 @@
-import { User, Login, Order, Address, ShippingPriceResponse, OrderToPlace } from '../shared/shareddtypes';
+import { User, Login, Order } from '../shared/shareddtypes';
 import {Product} from '../shared/shareddtypes';
 
 // export async function addUser(user:User):Promise<boolean>{
@@ -25,7 +25,7 @@ export async function getUsers():Promise<User[]>{
 
 export async function getProducts():Promise<Product[]>{
   const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000'
-  let response = await fetch(apiEndPoint+'/product');
+  let response = await fetch(apiEndPoint+'/products');
   //The objects returned by the api are directly convertible to Product objects
   return response.json()
 }
@@ -49,40 +49,6 @@ export async function getOrdersByUser(user_id:string):Promise<Order[]>{
   let response = await fetch(apiEndPoint+'/orders/'+user_id);
   //The objects returned by the api are directly convertible to Product objects
   return response.json()
-}
-
-export async function getShippingPrice(address: Address):Promise<ShippingPriceResponse>{
-  const requestOptions = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(
-      { number: address.number , 
-        street: address.street, 
-        city: address.city, 
-        country: address.country,
-        zipcode: address.zipcode})
-  };
-  const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000'
-  let response = await fetch(apiEndPoint+'/orders/price', requestOptions);
-  return response.json()
-}
-
-
-export async function placeOrder(orderToPlace:OrderToPlace):Promise<string>{
-  console.log("Entrando en llamada a place order");
-  const requestOptions = {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(orderToPlace)
-  };
-  const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000'
-  let response = await fetch(apiEndPoint+'/orders/add', requestOptions);
-
-  console.log("Saliendo de la llamada de place order");
-  console.log("Estado " + response.status)
-  let result = response.text()
-  console.log("Texto" + result)
-  return result
 }
