+++ webapp/src/components/Navbar/Navbar.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import ShoppingCartDrawer from '../Cart/ShoppingCartDrawer'
+import ShoppingCart from '../Cart/ShoppingCart'
 import { useSession } from "@inrupt/solid-ui-react";
 
 
@@ -31,8 +31,8 @@ const NavBar = () => {
   const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
     null,
   )
-  const { session, logout } = useSession();
-  const [message, setMessage] = useState(null as string | null);
+  const {session, logout} = useSession();
+  const [message, setMessage] = useState(null as string|null);
 
   const navigate = useNavigate()
 
@@ -54,9 +54,10 @@ const NavBar = () => {
   const handleLogout = () => {
     logout();
     handleCloseUserMenu();
-    navigate('/home')
-    setMessage('Successful logout')
+    setMessage('Sesión cerrada satisfactoriamente')
   }
+
+  console.log(session.info)
 
   return (
     <>
@@ -69,100 +70,100 @@ const NavBar = () => {
           {message}
         </Alert>
       </Snackbar>
-      <AppBar position="static" style={{ background: '#365073' }}>
-        <Container>
-          <Toolbar disableGutters>
-            <Typography
-              variant="h6"
-              noWrap
-              component="div"
-              sx={{ mr: 2, display: { xs: 'none', md: 'flex' } }}
+    <AppBar position="static" style={{ background: '#365073' }}>
+      <Container>
+        <Toolbar disableGutters>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ mr: 2, display: { xs: 'none', md: 'flex' } }}
+          >
+            <IconButton
+              size="large"
+              onClick={() => navigate('/')}
+              edge="start"
+              color="inherit"
+              aria-label="open drawer"
+              sx={{ mr: 4 }}
             >
-              <IconButton
-                size="large"
-                onClick={() => navigate('/')}
-                edge="start"
-                color="inherit"
-                aria-label="open drawer"
-                sx={{ mr: 4 }}
-              >
-                <img
-                  height="100px"
-                  src="/images/logo.jpg"
-                  alt="logo"
-                  className="logo"
-                />
-              </IconButton>
-            </Typography>
-            <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
-              <IconButton
-                size="large"
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleOpenNavMenu}
-                color="inherit"
-              >
-                <MenuIcon />
-              </IconButton>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorElNav}
-                anchorOrigin={{
-                  vertical: 'bottom',
-                  horizontal: 'left',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'left',
-                }}
-                open={Boolean(anchorElNav)}
-                onClose={handleCloseNavMenu}
-                sx={{
-                  display: { xs: 'block', md: 'none' },
-                }}
-              >
-                {pages.map((page) => (
-                  <MenuItem key={page} onClick={handleCloseNavMenu}>
-                    <Typography textAlign="center">{page}</Typography>
-                  </MenuItem>
-                ))}
-              </Menu>
-            </Box>
-            <Typography
-              variant="h6"
-              noWrap
-              component="div"
-              sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}
+              <img
+                height="100px"
+                src="/images/logo.jpg"
+                alt="logo"
+                className="logo"
+              />
+            </IconButton>
+          </Typography>
+          <Box sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}>
+            <IconButton
+              size="large"
+              aria-label="account of current user"
+              aria-controls="menu-appbar"
+              aria-haspopup="true"
+              onClick={handleOpenNavMenu}
+              color="inherit"
             >
-              <IconButton
-                size="large"
-                onClick={() => navigate('/')}
-                edge="start"
-                color="inherit"
-                aria-label="open drawer"
-                sx={{ mr: 4 }}
-              >
-                <img
-                  height="100px"
-                  src="/images/logo.jpg"
-                  alt="logo"
-                  className="logo"
-                />
-              </IconButton>
-            </Typography>
-            <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+              <MenuIcon />
+            </IconButton>
+            <Menu
+              id="menu-appbar"
+              anchorEl={anchorElNav}
+              anchorOrigin={{
+                vertical: 'bottom',
+                horizontal: 'left',
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: 'top',
+                horizontal: 'left',
+              }}
+              open={Boolean(anchorElNav)}
+              onClose={handleCloseNavMenu}
+              sx={{
+                display: { xs: 'block', md: 'none' },
+              }}
+            >
               {pages.map((page) => (
-                <Button
-                  key={page}
-                  onClick={handleCloseNavMenu}
-                  sx={{ my: 2, color: 'white', display: 'block' }}
-                >
-                  {page}
-                </Button>
+                <MenuItem key={page} onClick={handleCloseNavMenu}>
+                  <Typography textAlign="center">{page}</Typography>
+                </MenuItem>
               ))}
-            </Box>
+            </Menu>
+          </Box>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ flexGrow: 1, display: { xs: 'flex', md: 'none' } }}
+          >
+            <IconButton
+              size="large"
+              onClick={() => navigate('/')}
+              edge="start"
+              color="inherit"
+              aria-label="open drawer"
+              sx={{ mr: 4 }}
+            >
+              <img
+                height="100px"
+                src="/images/logo.jpg"
+                alt="logo"
+                className="logo"
+              />
+            </IconButton>
+          </Typography>
+          <Box sx={{ flexGrow: 1, display: { xs: 'none', md: 'flex' } }}>
+            {pages.map((page) => (
+              <Button
+                key={page}
+                onClick={handleCloseNavMenu}
+                sx={{ my: 2, color: 'white', display: 'block' }}
+              >
+                {page}
+              </Button>
+            ))}
+          </Box>
 
           <Box sx={{ flexGrow: 0 }}>
             <Tooltip title="Open settings">
@@ -203,7 +204,7 @@ const NavBar = () => {
               )}
             </Menu>
           </Box>
-          <ShoppingCartDrawer />
+          <ShoppingCart />
         </Toolbar>
       </Container>
     </AppBar>
